+++ flytekit/annotated/base_task.py
@@ -12,8 +12,8 @@ from flytekit.annotated.context_manager import (
     RegistrationSettings,
 )
 from flytekit.annotated.interface import Interface, transform_interface_to_typed_interface
-from flytekit.annotated.node import create_and_link_node
-from flytekit.annotated.promise import Promise, VoidPromise, create_task_output, translate_inputs_to_literals
+from flytekit.annotated.promise import Promise, VoidPromise, create_task_output, translate_inputs_to_literals, \
+    create_and_link_node
 from flytekit.annotated.type_engine import TypeEngine
 from flytekit.common.exceptions import user as _user_exceptions
 from flytekit.common.tasks.sdk_runnable import ExecutionParameters
