+++ flytekit/annotated/node.py
@@ -1,15 +1,9 @@
 from __future__ import annotations
 
-import datetime
-from typing import Any, List, Optional
+from typing import Any, List
 
-from flytekit.annotated import interface as flyte_interface
-from flytekit.annotated.context_manager import FlyteContext
-from flytekit.annotated.promise import Promise, VoidPromise, binding_from_python_std, create_task_output
 from flytekit.common import constants as _common_constants
-from flytekit.common.exceptions import user as _user_exceptions
 from flytekit.common.nodes import SdkNode
-from flytekit.common.promise import NodeOutput as _NodeOutput
 from flytekit.common.utils import _dnsify
 from flytekit.models import literals as _literal_models
 from flytekit.models.core import workflow as _workflow_model
@@ -115,7 +109,7 @@ class Node(object):
         return self._bindings
 
     @property
-    def upstream_nodes(self) -> List["Node"]:
+    def upstream_nodes(self) -> List[Node]:
         return self._upstream_nodes
 
     def with_overrides(self, *args, **kwargs):
@@ -128,93 +122,3 @@ class Node(object):
             self._aliases = []
             for k, v in alias_dict.items():
                 self._aliases.append(_workflow_model.Alias(var=k, alias=v))
-
-
-# TODO we should accept TaskMetadata here and then extract whatever fields we want into NodeMetadata
-def create_and_link_node(
-    ctx: FlyteContext,
-    entity,
-    interface: flyte_interface.Interface,
-    *args,
-    timeout: Optional[datetime.timedelta] = None,
-    retry_strategy: Optional[_literal_models.RetryStrategy] = None,
-    **kwargs,
-):
-    """
-    This method is used to generate a node with bindings. This is not used in the execution path.
-    """
-    if ctx.compilation_state is None:
-        raise _user_exceptions.FlyteAssertion("Cannot create node when not compiling...")
-
-    used_inputs = set()
-    bindings = []
-
-    typed_interface = flyte_interface.transform_interface_to_typed_interface(interface)
-
-    for k in sorted(interface.inputs):
-        var = typed_interface.inputs[k]
-        if k not in kwargs:
-            raise _user_exceptions.FlyteAssertion("Input was not specified for: {} of type {}".format(k, var.type))
-        v = kwargs[k]
-        # This check ensures that tuples are not passed into a function, as tuples are not supported by Flyte
-        # Usually a Tuple will indicate that multiple outputs from a previous task were accidentally passed
-        # into the function.
-        if isinstance(v, tuple):
-            raise AssertionError(
-                f"Variable({k}) for function({entity.name}) cannot receive a multi-valued tuple {v}."
-                f" Check if the predecessor function returning more than one value?"
-            )
-        bindings.append(
-            binding_from_python_std(
-                ctx, var_name=k, expected_literal_type=var.type, t_value=v, t_value_type=interface.inputs[k]
-            )
-        )
-        used_inputs.add(k)
-
-    extra_inputs = used_inputs ^ set(kwargs.keys())
-    if len(extra_inputs) > 0:
-        raise _user_exceptions.FlyteAssertion(
-            "Too many inputs were specified for the interface.  Extra inputs were: {}".format(extra_inputs)
-        )
-
-    # Detect upstream nodes
-    # These will be our annotated Nodes until we can amend the Promise to use NodeOutputs that reference our Nodes
-    upstream_nodes = list(
-        set(
-            [
-                input_val.ref.sdk_node
-                for input_val in kwargs.values()
-                if isinstance(input_val, Promise) and input_val.ref.node_id != _common_constants.GLOBAL_INPUT_NODE_ID
-            ]
-        )
-    )
-
-    node_metadata = _workflow_model.NodeMetadata(
-        f"{entity.__module__}.{entity.name}",
-        timeout or datetime.timedelta(),
-        retry_strategy or _literal_models.RetryStrategy(0),
-    )
-
-    # TODO: Clean up NodeOutput dependency on SdkNode, then rename variable
-    non_sdk_node = Node(
-        # TODO: Better naming, probably a derivative of the function name.
-        id=f"{ctx.compilation_state.prefix}n{len(ctx.compilation_state.nodes)}",
-        metadata=node_metadata,
-        bindings=sorted(bindings, key=lambda b: b.var),
-        upstream_nodes=upstream_nodes,  # type: ignore
-        flyte_entity=entity,
-    )
-    ctx.compilation_state.add_node(non_sdk_node)
-
-    if len(typed_interface.outputs) == 0:
-        return VoidPromise(entity.name)
-
-    # Create a node output object for each output, they should all point to this node of course.
-    node_outputs = []
-    for output_name, output_var_model in typed_interface.outputs.items():
-        # TODO: If node id gets updated later, we have to make sure to update the NodeOutput model's ID, which
-        #  is currently just a static str
-        node_outputs.append(Promise(output_name, _NodeOutput(sdk_node=non_sdk_node, sdk_type=None, var=output_name)))
-        # Don't print this, it'll crash cuz sdk_node._upstream_node_ids might be None, but idl code will break
-
-    return create_task_output(node_outputs, interface)
