+++ flytekit/annotated/workflow.py
@@ -16,13 +16,12 @@ from flytekit.annotated.interface import (
     transform_interface_to_typed_interface,
     transform_signature_to_interface,
 )
-from flytekit.annotated.node import Node, create_and_link_node
-from flytekit.annotated.promise import Promise, VoidPromise, create_task_output
+from flytekit.annotated.node import Node
+from flytekit.annotated.promise import Promise, VoidPromise, create_task_output, NodeOutput, create_and_link_node
 from flytekit.annotated.reference_entity import ReferenceEntity, WorkflowReference
 from flytekit.annotated.type_engine import TypeEngine
 from flytekit.common import constants as _common_constants
 from flytekit.common.exceptions.user import FlyteValidationException
-from flytekit.common.promise import NodeOutput as _NodeOutput
 from flytekit.common.workflow import SdkWorkflow as _SdkWorkflow
 from flytekit.loggers import logger
 from flytekit.models import interface as _interface_models
@@ -117,7 +116,7 @@ def _workflow_fn_outputs_to_promise(
 def construct_input_promises(inputs: List[str]):
     return {
         input_name: Promise(
-            var=input_name, val=_NodeOutput(sdk_node=GLOBAL_START_NODE, sdk_type=None, var=input_name,),
+            var=input_name, val=NodeOutput(node=GLOBAL_START_NODE, sdk_type=None, var=input_name),
         )
         for input_name in inputs
     }
